requirements.txt
amqp==5.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.0
async-timeout==4.0.3
billiard==4.2.0
boto3==1.34.79
botocore==1.34.79
Brotli==1.1.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cssselect2==0.7.0
dj-database-url==2.1.0
Django==4.2.11
django-dump-load-utf8==0.0.4
django-filter==24.2
django-minio-storage==0.5.7
django-storages==1.14.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
fonttools==4.51.0
html5lib==1.1
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.6
minio==7.2.5
packaging==24.0
pillow==10.3.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.20.0
pydyf==0.9.0
PyJWT==2.8.0
pyphen==0.14.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
tinycss2==1.2.1
typing-extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
weasyprint==61.2
webencodings==0.5.1
zopfli==0.2.3
